Fix prohibit_public_ip_on_vnic
Imports subnet's prohibit_public_ip_on_vnic property and forces
new when changed. In addition to the test update, also validated
that if a statefile and plan does not have this property set
(which would be the case for plans run before this property
existed), then upgrading and running plan will correctly compare the
value with the default and make no changes.