Sync instance subnet and hostname_label
An instances have 2 ways to set subnet_id and hostname_label: either at the
root level, or in create_vnic_details. By syncing both sets of properties with
the VNIC and setting all of these to computed, this change allows users to
switch between which version is used without causing changes to their plan.
This was partly in place before, but this change puts a few missing pieces in
place.