Add solutions folder
The docs/examples folder currently contains a mix of low-level examples and
very high level "how to complete scenario X" examples. This change starts the
migration of that second cagetory into a new section under docs/solutions.
(There are a few other examples that should probably be moved as well, but I need
to get signoff from the owners first.)
This also adds a placeholder for an upcoming SQL Server solution.