Add source port range to security list rules (another rebase) (#385)
Ported from PR #366.

Currently, TCP and UDP options can be specified with a single set of
max/min values, which correspond to the destination port range. This change
adds the ability to specify the source port range as well, issue #338.

Ideally, we should deprecate the existing max/min params and add a new
destination_port_range option next to the new source_port_range. Unfortunately,
a number of issue with nested schemas in Terraform are preventing support of both
of these params simultaneously. Instead, max/min params will remain, and source_port_range
will exist beside it.

This change also enables security lists with no rules.

Some updates are made to the docs, but the docs for the security list resource and data source
could really use a bigger update that is not included in this change.