Make examples testable
Adds test cases TestExamplesPlan and TestExamplesApply, which will plan
and apply every example under /docs/examples. (There's an exception for
db_systems, which needs some work to make runnable using the stardard
set of variables.) The tests ensure that all examples are runnable using
the same set of variables.

In the future we might consider updating this to enable concurrent runs.