Update RHEL solution to use image OCID (#399)
The image "Oracle-Linux-7.4-2017.09.29-0" is no longer being returned from the list_images call, so
use the OCID directly (which is the new guidance anyway, we just haven't yet switched all of the
solutions and examples over). Note that we might want to update this to a newer image, such as
"Oracle-Linux-7.4-2017.11.15-0" or "Oracle-Linux-7.4-2017.12.18-0", but that is pending testing using
those versions.