RHEL solution Enhancement requests... (#407)
This update accomplishes several items:

More fully documents requirements for subnet security lists, AD specification format and exclusion of root compartment from use by this process.
Includes inline documentation in the configuration.tf file
Changes access method of RHEL ISO from OCI bucket/object method to URL. This allows any valid URL that contains the name of the ISO to be used in this process.
Changed the method of image selection for the ipxe instance to use the Operating System and OS Version vs. specifying the image name directly. Allows for selection of current images and ensures that deprecated images are not used.
Fixes issue when using on certain Linux distros where the fingerprint generation inserts an extraneous '(stdin)'
Removes the need for shar in generating the user-data script.