Comment out "domain" attribute as optional in the DBSystem example
When a subnet is created, Internet and VCN resolver is enabled by default.
In this case, "domain" attribute shouldn't be provided otherwise you get
an error if the attribute matches the domain name configured for the subnet.
It is reasonable not to provide "domain" attribute in this example because
when customers try this example, they most likely use a subnet with the
default DNS settings.  If a customer is strongly aware of the DNS settings,
he/she would be able to provide the right value for "domain" attribute.