Add mutex to synchronize lb backend and backend_set operations.
This introduces a per-backend_set mutex to serialize concurrent accesses that modify a
backend_set either via a change in the backend or backend_set resource.

The Lock/Unlock mechanism is added to generic CRUD operations. Any resource
can implement an interface that returns a mutex to be used for synchronization.

Addresses issue found in github #519