Updated with fixes and image OCID updates (#538)
What is fixed/changed in this release:

Fixed issue with OCI-CLI and pip that was preventing the process from working. The original
process attempted to install OCI-CLI as a global resource in the global python repo...this broke
after some changes in pip. So now we install pip locally to each user.
Changed our blank image OCID to use a new global value. This makes the process more portable across
regions (but not completely...see next bullet).
Updated the ipxe server image list to include the LHR region. LHR is now fully supported on this process.
*** IMPORTANT CHANGE *** Updated process to use Instance Principal authorization method for OCI-CLI. This
removes the requirement to upload the user private key into the ipxe server. HOWEVER, it does REQUIRE
the creation of a Dynamic Group for the Compartment this is being executed in. This is included in the
prerequisites section below. The new process WILL NOT WORK without an Instance Principal (also known
as a Dynamic Group).