Workaround a race condition in the IPSecConnections force new test
This is to workaround an issue where Terraform doesn't delete resources in order
when it is a part of a ForceNew update.

The issue happens when IPSecConnection resource is created with a dependency on CPE "A". We then create
a CPE "B", remove CPE "A", and point IPSecConnection to CPE "B" in a subsequent apply.

This should force a new IPSecConnection and the expected order of operations was:
1) Delete IPSecConnection
2) Delete CPE "A" (This must be deleted after IPSecConnection due to a dependency on this)
3) Create CPE "B"
4) Create new IPSecConnection dependent on CPE "B"

The observed order of operations in TF was:
1) Delete CPE "A" (this results in a 409 error because IPSecConnection hasn't been removed first)
2) Delete IPSecConnection

Step 1 and 2 happened concurrently because TF does not detect a dependency between the two resources in
the ForceNew case.

Fortunately, this is not an issue with the Destroy case.

We avoid this issue by keeping both CPE "A" and "B" in the test.