Fix private ip test to avoid assumptions about ordering
The datasource test verifies that we get 2 private IPs back for a vnic
from 1 instance. The test was assuming that the first private IP is always
the primary (and hence was checking for the hostname label).

Fix the test to check for the hostname label in either returned private ip.