Reorder DHCP options test steps to avoid leaking resources
With the dependency on v0.11.7 of Terraform dependencies, an issue was introduced
in tests where a config that is eminently invalid would result in the state being
obliterated.

Some tests may have an ExpectError step at the end of a test case that may cause the valid
state from a previous step to be obliterated. This ends up leaving behind resources that
are unmanaged by tests. When the test case tries to clean up the resources, it is not able
to find any to clean up.

This issue was causing a leak in VCNs in our tests.

Upon review of tests; only this one appears to have this behavior.