Fix load balance rule set test to verify datasource after dependency change
In v0.12 Terraform, there are fixes to ensure refresh of datasources occur if a
dependency has been updated. The test was erroneously taking an assumption that no
refresh would happen and was using dependency from a previous test step during validation.

Fix this so the datasource validation reflects the state of the current dependency.