replace deprecated terraform instrumentation pkgs
The packages hashicorp/terraform/backend/local and hashicorp/terraform/command are deprecated for Terraform providers and are not available in the standalone plugin SDK module, hashicorp/terraform-plugin-sdk.

This PR removes these imports and adds a new runTerraform(args...) function for instrumenting the terraform CLI command. An error will be returned if this command cannot be found in the system PATH.