[Resource Discovery] Fix issue where singular datasources result in success even though nothing was found during discovery.
This issue was caused by the `handleMissingResourceError` that is used to nullify missing resource errors (404).
This resulted in resource discovery assuming that discovery succeeded for singular datasources and proceeding to import a non-existent resource.

This fix checks that singular data source results are not voided before proceeding to add it to discovered resource list. Voided singular data source results are considered a case of resource not found.