Avoid chunked HTTP requests for Object Storage
In go net/http, the recommend way to signal an empty body
is using http.NoBody. (or leaving the body nil.) With this
patch, for empty strings and empty body cases, we use
http.NoBody explicitly to avoid any accidental chunked
transfer encoding in HTTP requests.

Before this change, the issue can be identified by the presence
of "Transfer-Encoding: chunked" header and absence of "Content-Length"
header in the requests.

Signed-off-by: Tolga Ceylan <tolga.ceylan@oracle.com>