Fix ExaCS cloud VM cluster as source of `oci_database_db_home` resource creation
Remove polling to wait for resource to reach non-updating state in db home. This polling logic was bug prone and wasn't doing what was intended. It also uses `GetVmCluster` API to poll status of a vm cluster regardless of type, which does not work for ExaCS clusters. This would often result in errors from DB service about using the proper APIs.

Instead, let's just keep retrying the update function if it fails. This is a more general solution to resource creation failures.